requirements.txt
python-dotenv
streamlit
audio_recorder_streamlit
torch
pyaudio
whisper
wave
stqdm
